# backend/book-social-network/src/main/java/com/kekkobarr/booksocialnetwork/auth/AuthenticationResponse.java
package com.kekkobarr.booksocialnetwork.auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {

    private String token;
}