# backend/book-social-network/src/main/java/com/kekkobarr/booksocialnetwork/feedback/FeedbackResponse.java
package com.kekkobarr.booksocialnetwork.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}